services/django-api/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
boto3==1.34.0
psycopg-binary==2.9.9
redis==5.0.1
celery==5.3.4
kombu==5.3.4
django-cors-headers==4.3.1
gunicorn==21.2.0
python-dotenv==1.0.0
cryptography==41.0.7

